chore: update to ESLint v9
ESLint has been updated to v9 and the configuration converted to the new
"flat" config format. Related packages were updated to versions that
support ESLint v9.

- `.eslintrc.js` becomes `eslint.config.mjs`, and `.eslintignore`
  becomes the global `ignores` entry. The config has to be an ES module
  because `eslint-plugin-import-x` ships separate CommonJS and ESM
  builds, and the plugin object registered here has to be identical to
  the one `@metamask/eslint-config-typescript` registers.
- `tests/framework/.eslintrc.js` becomes `tests/framework/eslint-config.js`,
  with each glob rooted at `tests/` since `extends` inside an override
  used to intersect the outer and inner file globs.
- `@react-native/eslint-config` is still eslintrc-only, so it is
  translated with `FlatCompat`.
- Flat config no longer skips dot-directories, so `.github`,
  `.storybook`, and `.yarn` are ignored explicitly to keep the linted
  file set unchanged.
- A few rules newly enabled by the dependency updates are turned off for
  now: jsdoc's `recommended-typescript-error` preset, and import-x's
  `no-rename-default` and `no-named-as-default-member`.

`yarn lint` reports the same 0 errors as before, and the same warnings
apart from the four `import-x/no-named-as-default-member` ones.
