--- .eslintignore
@@ -1,19 +1,0 @@
-/scripts/inpage-bridge
-/app/core/InpageBridgeWeb3.js
-/app/components/UI/Charts/AdvancedChart/webview/chartLogic.js
-/app/components/UI/Charts/AdvancedChart/webview/chartLogicString.ts
-/app/components/UI/Charts/AdvancedChart/webview/syncChartLogic.js
-/app/util/blockies.js
-__snapshots__
-android
-build
-coverage
-ios
-jest.preprocessor.js
-node_modules
-CHANGELOG.md
-/app/util/test/testSetup.js
-junitProperties.js
-/tests/scripts/ai-e2e-tags-selector.ts
-tests/scripts/aggregate-performance-reports.mjs
-app/**/*-method-action-types.ts
