+++ metro.transform.js
@@ -176,46 +176,28 @@ let eslintInstance;
  */
 function getESLintInstance() {
   if (!eslintInstance) {
-    const eslintrc = require('./.eslintrc.js');
+    const flatConfig = require('./eslint.config.mjs').default;
 
-    eslintrc.overrides.forEach((override) => {
-      const rules = override.rules ?? {};
+    const baseConfig = [
+      // Drop the test-only config entries. We will never lint test files here.
+      ...flatConfig.filter(
+        (entry) => !(entry.files && entry.plugins && 'jest' in entry.plugins),
+      ),
+      {
+        rules: {
+          // We don't want linting to fail for purely stylistic reasons.
+          'prettier/prettier': 'off',
+          // Sometimes we use `let` instead of `const` to assign variables depending on
+          // the build type.
+          'prefer-const': 'off',
+          // This rule is broken when run under lockdown with `process.env.CI` set to `true`.
+          // Temporarily disabled until we can eliminate code fences.
+          'react/no-unescaped-entities': 'off',
+        },
+      },
+    ];
 
-      // We don't want linting to fail for purely stylistic reasons.
-      rules['prettier/prettier'] = 'off';
-      // Sometimes we use `let` instead of `const` to assign variables depending on
-      // the build type.
-      rules['prefer-const'] = 'off';
-
-      override.rules = rules;
-    });
-
-    // also override the rules section
-    // We don't want linting to fail for purely stylistic reasons.
-    eslintrc.rules['prettier/prettier'] = 0;
-    // Sometimes we use `let` instead of `const` to assign variables depending on
-    // the build type.
-    eslintrc.rules['prefer-const'] = 0;
-
-    // This rule is broken when run under lockdown with `process.env.CI` set to `true`.
-    // Temporarily disabled until we can eliminate code fences.
-    eslintrc.rules['react/no-unescaped-entities'] = 0;
-
-    // Remove all test-related overrides. We will never lint test files here.
-    eslintrc.overrides = eslintrc.overrides.filter(
-      (override) =>
-        !(
-          (override.extends &&
-            override.extends.find(
-              (configName) =>
-                configName.includes('jest') || configName.includes('mocha'),
-            )) ||
-          (override.plugins &&
-            override.plugins.find((pluginName) => pluginName.includes('jest')))
-        ),
-    );
-
-    eslintInstance = new ESLint({ baseConfig: eslintrc, useEslintrc: false });
+    eslintInstance = new ESLint({ baseConfig, overrideConfigFile: true });
   }
   return eslintInstance;
 }
