+++ package.json
@@ -590,6 +590,8 @@
     "@babel/preset-env": "^7.25.3",
     "@babel/register": "^7.24.6",
     "@babel/runtime": "^7.25.0",
+    "@eslint/eslintrc": "^3.3.1",
+    "@eslint/js": "^9.39.2",
     "@ethersproject/contracts": "^5.7.0",
     "@ethersproject/providers": "^5.7.2",
     "@expo/metro-config": "~55.0.23",
@@ -602,7 +604,8 @@
     "@metamask/build-utils": "^3.0.0",
     "@metamask/client-mcp-core": "^0.8.0",
     "@metamask/device-mcp": "^0.3.0",
-    "@metamask/eslint-config-typescript": "^13.0.0",
+    "@metamask/eslint-config": "^14.1.1",
+    "@metamask/eslint-config-typescript": "^14.1.1",
     "@metamask/eslint-plugin-design-tokens": "^1.0.0",
     "@metamask/foundryup": "1.0.0",
     "@metamask/messenger-cli": "^0.2.0",
@@ -658,8 +661,8 @@
     "@types/url-parse": "^1.4.8",
     "@types/valid-url": "^1.0.4",
     "@types/ws": "^8.18.1",
-    "@typescript-eslint/eslint-plugin": "^8.1.0",
-    "@typescript-eslint/parser": "^8.1.0",
+    "@typescript-eslint/eslint-plugin": "^8.67.0",
+    "@typescript-eslint/parser": "^8.67.0",
     "@walletconnect/types": "^2.23.0",
     "@wdio/protocols": "^9.27.0",
     "@welldone-software/why-did-you-render": "^8.0.1",
@@ -688,20 +691,21 @@
     "eas-cli": "^12.6.1",
     "esbuild": "^0.25.10",
     "esbuild-register": "^3.6.0",
-    "eslint": "^8.44.0",
-    "eslint-config-prettier": "^8.1.0",
+    "eslint": "^9.39.2",
+    "eslint-config-prettier": "^10.1.5",
     "eslint-import-resolver-typescript": "^3.6.3",
-    "eslint-plugin-import-x": "^0.5.1",
+    "eslint-plugin-import-x": "^4.3.0",
     "eslint-plugin-jest": "^29.14.0",
-    "eslint-plugin-jsdoc": "43.0.7",
-    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-jsdoc": "^50.2.4",
+    "eslint-plugin-prettier": "^5.5.1",
     "eslint-plugin-promise": "^7.2.1",
-    "eslint-plugin-react": "7.34.1",
+    "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-compiler": "^19.1.0-rc.2",
-    "eslint-plugin-react-native": "^4.0.0",
+    "eslint-plugin-react-native": "^5.0.0",
     "eslint-plugin-tailwindcss": "^3.18.2",
     "execa": "^8.0.1",
     "ganache": "^7.9.2",
+    "globals": "^15.15.0",
     "husky": "^9.1.7",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
@@ -740,6 +744,7 @@
     "ts-node": "^10.9.2",
     "tsx": "^4.21.0",
     "typescript": "~5.4.5",
+    "typescript-eslint": "^8.67.0",
     "webdriverio": "^9.27.0",
     "webextension-polyfill": "^0.12.0",
     "webpack": "^5.88.2",
